// src/api/entities/tutanota/TypeRefs.ts
export const ConversationEntryTypeRef: TypeRef<ConversationEntry> = new TypeRef("tutanota", "ConversationEntry")

export function createConversationEntry(values?: Partial<ConversationEntry>): ConversationEntry {
	return Object.assign(create(typeModels.ConversationEntry, ConversationEntryTypeRef), values)
}

export type ConversationEntry = {
	_type: TypeRef<ConversationEntry>;

	_format: NumberString;
	_id: IdTuple;
	_ownerGroup: null | Id;
	_permissions: Id;
	conversationType: NumberString;
	messageId: string;

	mail:  null | IdTuple;
	previous:  null | IdTuple;
}
// ...
export type EncryptedMailAddress = {
	_type: TypeRef<EncryptedMailAddress>;

	_id: Id;
	address: string;
	name: string;
}
// ...
export const MailTypeRef: TypeRef<Mail> = new TypeRef("tutanota", "Mail")

export function createMail(values?: Partial<Mail>): Mail {
	return Object.assign(create(typeModels.Mail, MailTypeRef), values)
}

export type Mail = {
	_type: TypeRef<Mail>;
	_errors: Object;

	_format: NumberString;
	_id: IdTuple;
	_ownerEncSessionKey: null | Uint8Array;
	_ownerGroup: null | Id;
	_permissions: Id;
	authStatus: null | NumberString;
	confidential: boolean;
	differentEnvelopeSender: null | string;
	listUnsubscribe: boolean;
	method: NumberString;
	movedTime: null | Date;
	phishingStatus: NumberString;
	receivedDate: Date;
	recipientCount: NumberString;
	replyType: NumberString;
	sentDate: null | Date;
	state: NumberString;
	subject: string;
	unread: boolean;

	attachments: IdTuple[];
	bccRecipients: MailAddress[];
	body:  null | Id;
	bucketKey:  null | BucketKey;
	ccRecipients: MailAddress[];
	conversationEntry: IdTuple;
	firstRecipient:  null | MailAddress;
	headers:  null | Id;
	mailDetails:  null | IdTuple;
	mailDetailsDraft:  null | IdTuple;
	replyTos: EncryptedMailAddress[];
	restrictions:  null | MailRestriction;
	sender: MailAddress;
	toRecipients: MailAddress[];
}
export const MailAddressTypeRef: TypeRef<MailAddress> = new TypeRef("tutanota", "MailAddress")

export function createMailAddress(values?: Partial<MailAddress>): MailAddress {
	return Object.assign(create(typeModels.MailAddress, MailAddressTypeRef), values)
}

export type MailAddress = {
	_type: TypeRef<MailAddress>;

	_id: Id;
	address: string;
	name: string;

	contact:  null | IdTuple;
}
// ...
export type MailRestriction = {
	_type: TypeRef<MailRestriction>;

	_id: Id;

	delegationGroups_removed: Id[];
	participantGroupInfos: IdTuple[];
}